+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### v0.9.0
 
  - adds Break(key,value,...) for temporary runtime inspection. 
+ - add optional filename to Dump()
 
 ### v0.8.1